requirements.txt
Flask
playwright
python-dotenv
beautifulsoup4